fix: mise a jour recette ml
